spaced-repetition/backend/requirements.txt
# Django and Core dependencies
Django==5.1.7
djangorestframework==3.16.0
django-cors-headers==4.7.0
asgiref==3.8.1
sqlparse==0.5.3
pytz==2025.2

# Database dependencies
psycopg2-binary==2.9.9  # PostgreSQL adapter
djongo==1.2.31
pymongo==4.11.3
dnspython==2.7.0
dataclasses==0.6

# Production dependencies
gunicorn==21.2.0  # WSGI HTTP Server
whitenoise==6.6.0  # Static file serving
dj-database-url==2.1.0  # Database URL configuration

# AWS specific
boto3==1.33.0  # For AWS S3 integration
django-storages==1.14.2  # Storage backends

# Security
django-environ==0.11.2  # Environment variable handling
python-dotenv==1.0.0  # Load environment variables from .env file

# Social Authentication
django-allauth==0.57.0  # Social authentication

# Optional for improved performance
uvicorn==0.27.1  # ASGI server (alternative to gunicorn)

# Additional production dependencies
requests==2.32.4  # HTTP library
certifi==2025.7.14  # SSL certificates

# AI and LangGraph dependencies
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-openai>=0.1.0
langgraph==0.2.28
langsmith>=0.1.0
google-generativeai==0.7.2
python-json-logger>=2.0.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.2
Pillow==10.4.0  # For image processing

# Web scraping and content extraction
beautifulsoup4==4.12.3
newspaper3k==0.2.8
lxml==5.1.0
youtube-transcript-api>=1.2.0
Wikipedia-API==0.8.1
validators==0.22.0
urllib3==2.2.0
